feat: use`ForAttributeWithMetadataName` instead of `CreateSyntaxProvider`
