+++ src/Riok.Mapperly/MapperGenerator.cs
@@ -14,41 +14,15 @@ namespace Riok.Mapperly;
 public class MapperGenerator : IIncrementalGenerator
 {
     private const string GeneratedFileSuffix = ".g.cs";
-    private static readonly string _mapperAttributeName = typeof(MapperAttribute).FullName;
+
+    public static readonly string MapperAttributeName = typeof(MapperAttribute).FullName;
 
     public void Initialize(IncrementalGeneratorInitializationContext context)
     {
-        var mapperClassDeclarations = context.SyntaxProvider
-            .CreateSyntaxProvider(
-                static (s, _) => IsSyntaxTargetForGeneration(s),
-                static (ctx, _) => GetSemanticTargetForGeneration(ctx))
-            .WhereNotNull();
+        var mapperClassDeclarations = SyntaxProvider.GetClassDeclarations(context);
 
         var compilationAndMappers = context.CompilationProvider.Combine(mapperClassDeclarations.Collect());
         context.RegisterImplementationSourceOutput(compilationAndMappers, static (spc, source) => Execute(source.Left, source.Right, spc));
-    }
-
-    private static bool IsSyntaxTargetForGeneration(SyntaxNode node)
-        => node is ClassDeclarationSyntax { AttributeLists.Count: > 0 };
-
-    private static ClassDeclarationSyntax? GetSemanticTargetForGeneration(GeneratorSyntaxContext ctx)
-    {
-        var classDeclaration = (ClassDeclarationSyntax)ctx.Node;
-        foreach (var attributeListSyntax in classDeclaration.AttributeLists)
-        {
-            foreach (var attributeSyntax in attributeListSyntax.Attributes)
-            {
-                if (ctx.SemanticModel.GetSymbolInfo(attributeSyntax).Symbol is not IMethodSymbol attributeSymbol)
-                    continue;
-
-                var attributeContainingTypeSymbol = attributeSymbol.ContainingType;
-                var fullName = attributeContainingTypeSymbol.ToDisplayString();
-                if (fullName == _mapperAttributeName)
-                    return classDeclaration;
-            }
-        }
-
-        return null;
     }
 
     private static void Execute(Compilation compilation, ImmutableArray<ClassDeclarationSyntax> mappers, SourceProductionContext ctx)
@@ -59,8 +33,7 @@ public class MapperGenerator : IIncrementalGenerator
 #if DEBUG_SOURCE_GENERATOR
         DebuggerUtil.AttachDebugger();
 #endif
-
-        var mapperAttributeSymbol = compilation.GetTypeByMetadataName(_mapperAttributeName);
+        var mapperAttributeSymbol = compilation.GetTypeByMetadataName(MapperAttributeName);
         if (mapperAttributeSymbol == null)
             return;
