+++ src/Riok.Mapperly/SyntaxProvider.Roslyn4.0.cs
@@ -1,0 +1,43 @@
+﻿#if !ROSLYN4_4_OR_GREATER
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Riok.Mapperly.Helpers;
+
+namespace Riok.Mapperly;
+
+internal static class SyntaxProvider
+{
+    public static IncrementalValuesProvider<ClassDeclarationSyntax> GetClassDeclarations(IncrementalGeneratorInitializationContext context)
+    {
+        return context.SyntaxProvider
+            .CreateSyntaxProvider(
+                static (s, _) => IsSyntaxTargetForGeneration(s),
+                static (ctx, _) => GetSemanticTargetForGeneration(ctx))
+            .WhereNotNull();
+    }
+
+    private static bool IsSyntaxTargetForGeneration(SyntaxNode node)
+        => node is ClassDeclarationSyntax { AttributeLists.Count: > 0 };
+
+    private static ClassDeclarationSyntax? GetSemanticTargetForGeneration(GeneratorSyntaxContext ctx)
+    {
+        var classDeclaration = (ClassDeclarationSyntax)ctx.Node;
+        foreach (var attributeListSyntax in classDeclaration.AttributeLists)
+        {
+            foreach (var attributeSyntax in attributeListSyntax.Attributes)
+            {
+                if (ctx.SemanticModel.GetSymbolInfo(attributeSyntax).Symbol is not IMethodSymbol attributeSymbol)
+                    continue;
+
+                var attributeContainingTypeSymbol = attributeSymbol.ContainingType;
+                var fullName = attributeContainingTypeSymbol.ToDisplayString();
+                if (fullName == MapperGenerator.MapperAttributeName)
+                    return classDeclaration;
+            }
+        }
+
+        return null;
+    }
+}
+#endif
